requirements.txt
pyelftools==0.29
termcolor==2.4.0
